\name{bpmn-visualization-r-package}
\alias{bpmn-visualization-r-package}
\alias{bpmn-visualization-r}
\docType{package}
\title{
\packageTitle{bpmn-visualization-r}
}
\description{
\packageDescription{bpmn-visualization-r}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bpmn-visualization-r}
\packageIndices{bpmn-visualization-r}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bpmn-visualization-r}

Maintainer: \packageMaintainer{bpmn-visualization-r}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
