% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmnVisualizationR.R
\name{bpmnVisualizationR-shiny-output}
\alias{bpmnVisualizationR-shiny-output}
\alias{bpmnVisualizationROutput}
\title{Shiny output binding for the \code{bpmnVisualizationR} 'HTML' widget}
\usage{
bpmnVisualizationROutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}
}
\value{
An output function that enables the use of the \code{bpmnVisualizationR} widget within 'Shiny' applications.
}
\description{
Helper to create output function for using the \code{bpmnVisualizationR} 'HTML' widget within 'Shiny' applications and interactive 'Rmd' documents.
}
