% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_fill}
\alias{create_fill}
\title{Create the fill style of an overlay}
\usage{
create_fill(color)
}
\arguments{
\item{color}{The color of the background of the overlay}
}
\value{
The fill style object of the overlay
}
\description{
When adding an overlay to an existing element in a diagram, it's possible to customize how it is filled.

Refer to the \code{fill} parameter in the \code{\link{create_style}} function for more information.

Use this function to create the correct fill structure for an overlay.
}
