% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_font}
\alias{create_font}
\title{Create the font style of an overlay}
\usage{
create_font(color = NULL, size = NULL)
}
\arguments{
\item{color}{The color of the font of an overlay}

\item{size}{The size of the font of an overlay}
}
\value{
The font style object of an overlay
}
\description{
When an overlay is added to an existing element in the diagram, it's possible to style its font.

See \code{font} argument in the \code{\link{create_style}} function.

Use this structure to create correct font structure.
}
