% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_gradient_fill}
\alias{create_gradient_fill}
\title{Create a gradient fill style for an element}
\usage{
create_gradient_fill(direction, start_color, end_color)
}
\arguments{
\item{direction}{The direction of the gradient (e.g., \code{left-to-right}, \code{right-to-left}, \code{bottom-to-top}, \code{top-to-bottom}).
Taking the example of bottom-to-top, this means that the start color is at the bottom of the paint pattern and the end color is at the top, with a gradient between them.}

\item{start_color}{The starting color of the gradient. It can be any HTML color name or HEX code, as well as special keywords such as \code{inherit}, \code{none}, \code{swimlane}.}

\item{end_color}{The ending color of the gradient. It can be any HTML color name or HEX code, as well as special keywords such as \code{inherit}, \code{none}, \code{swimlane}.}
}
\value{
A gradient fill style object.
}
\description{
Create a gradient fill style for an element.
}
