% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_overlay}
\alias{create_overlay}
\title{Create an overlay}
\usage{
create_overlay(elementId, label, style = NULL, position = NULL)
}
\arguments{
\item{elementId}{The bpmn element id to which the overlay will be attached}

\item{label}{'HTML' element to use as an overlay}

\item{style}{The style of the overlay
Use \code{\link{create_style}} function to create the style object of an overlay.}

\item{position}{The position of the overlay
If the bpmn element where the overlay will be attached is a Shape, use \code{\link{OverlayShapePosition}}.
Otherwise, use \code{\link{OverlayEdgePosition}}.}
}
\value{
An overlay object
}
\description{
An overlay can be added to existing elements in the diagram.

See \code{overlays} argument in the \code{\link{display}} function.

Use this function to create the correct overlay structure.
}
\examples{
# Create an overlay with shape position "top-left"
overlay <- create_overlay(
  "my-element-id", 
  "My Overlay Label", 
  create_style(
    font = create_font(color = 'DarkSlateGray', size = 23),
    fill = create_fill(color = 'MistyRose'),
    stroke = create_stroke(color = 'Red')
  ),
  OverlayShapePosition[1]
)

# Create an overlay with edge position "end"
overlay <- create_overlay(
  "my-edge-id", 
  "My Overlay Label", 
  create_style(
    font = create_font(color = 'DarkSlateGray', size = 23),
    fill = create_fill(color = 'MistyRose'),
    stroke = create_stroke(color = 'Red')
  ),
  OverlayEdgePosition[2]
)

}
