% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_stroke}
\alias{create_stroke}
\title{Create the stroke style of an overlay}
\usage{
create_stroke(color)
}
\arguments{
\item{color}{The color of the stroke of an overlay}
}
\value{
The stroke style object of an overlay
}
\description{
When an overlay is added to an existing element in the diagram, it's possible to style its stroke.

See \code{stroke} argument in the \code{\link{create_style}} function.

Use this structure to create correct stroke structure.
}
