% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_style}
\alias{create_style}
\title{Create the style of an overlay}
\usage{
create_style(font = NULL, fill = NULL, stroke = NULL)
}
\arguments{
\item{font}{The font style of the overlay
Use \code{\link{create_font}} function to create the font style object for the overlay.}

\item{fill}{The fill style of the overlay
Use \code{\link{create_fill}} function to create the fill style object for the overlay.}

\item{stroke}{The stroke style of the overlay
Use \code{\link{create_stroke}} function to create the stroke style object for the overlay.}
}
\value{
The style object of the overlay
}
\description{
When adding an overlay to an existing element in a diagram, it's possible to customize its style.

Refer to the \code{style} parameter in the \code{\link{create_overlay}} function for more information.

Use this function to create the correct style structure for an overlay.
}
