% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\docType{data}
\name{overlay_edge_position}
\alias{overlay_edge_position}
\title{The overlay positions on \code{Edge}}
\format{
An object of class \code{character} of length 3.
}
\usage{
overlay_edge_position
}
\description{
To specify the position when creating an overlay object that will be attached to BPMN \code{Edge} elements in the diagram.

\code{overlay_edge_position} includes the following positions:
- \code{"start"}
- \code{"end"}
- \code{"middle"}

Use these constants as the \code{position} argument in the \code{\link{create_overlay}} function.
}
\keyword{datasets}
