% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\docType{data}
\name{overlay_shape_position}
\alias{overlay_shape_position}
\title{The overlay positions on \code{Shape}}
\format{
An object of class \code{character} of length 8.
}
\usage{
overlay_shape_position
}
\description{
To specify the position when creating an overlay object that will be attached to BPMN \code{Shape} elements in the diagram.

\code{overlay_shape_position} includes the following positions:

- \code{"top-left"}
- \code{"top-right"}
- \code{"top-center"}
- \code{"bottom-left"}
- \code{"bottom-right"}
- \code{"bottom-center"}
- \code{"middle-left"}
- \code{"middle-right"}

Use these constants as the \code{position} argument in the \code{\link{create_overlay}} function.
}
\keyword{datasets}
