\name{bpmnVisualization-package}
\alias{bpmnVisualization-package}
\alias{bpmnVisualization}
\docType{package}
\title{
\packageTitle{bpmnVisualization}
}
\description{
\packageDescription{bpmnVisualization}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bpmnVisualization}
\packageIndices{bpmnVisualization}
}
\author{
\packageAuthor{bpmnVisualization}

Maintainer: \packageMaintainer{bpmnVisualization}
}
\examples{
\dontrun{

bpmnVisualization::display(bpmn_xml)

}

}

\section{More examples}{


  See \url{https://github.com/process-analytics/bpmn-visualization-R#readme} for more usage examples.
}

\section{Support}{


  Use \url{https://github.com/process-analytics/bpmn-visualization-R/issues} for bug reports
}
