% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmnVisualization-package.R
\docType{package}
\name{bpmnVisualization-package}
\alias{bpmnVisualization}
\alias{bpmnVisualization-package}
\title{bpmnVisualization: Visualize Process Execution Data on BPMN Diagrams}
\description{
To visualize the execution data of the processes on BPMN (Business Process Model and Notation) diagrams, using overlays, style customization and interactions, with BPMN Visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/process-analytics/bpmn-visualization-R}
  \item Report bugs at \url{https://github.com/process-analytics/bpmn-visualization-R/issues}
}

}
\author{
\strong{Maintainer}: Celine Souchet \email{process.analytics.dev+CRAN@gmail.com}

Authors:
\itemize{
  \item Thomas Bouffard
}

}
\keyword{internal}
