\name{bpmnVisualization-package}
\alias{bpmnVisualization-package}
\alias{bpmnVisualization}
\docType{package}
\title{
\packageTitle{bpmnVisualization}
}
\description{
\packageDescription{bpmnVisualization}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bpmnVisualization}
\packageIndices{bpmnVisualization}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bpmnVisualization}

Maintainer: \packageMaintainer{bpmnVisualization}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ 
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
