% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmnVisualization.R
\name{renderBpmnVisualization}
\alias{renderBpmnVisualization}
\title{Shiny render binding for the \code{bpmn-visualization} HTML widget}
\usage{
renderBpmnVisualization(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a \code{bpmn-visualization} HTML widget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
A render function that enables the use of the \code{bpmn-visualization} widget within Shiny applications.
}
\description{
Helper to create render function for using the \code{bpmn-visualization} HTML widget within Shiny applications and interactive Rmd documents.
}
