% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmnVisualization.R
\name{bpmnVisualization-shiny}
\alias{bpmnVisualization-shiny}
\alias{bpmnVisualizationOutput}
\alias{renderBpmnVisualization}
\title{Shiny bindings for bpmnVisualization}
\usage{
bpmnVisualizationOutput(outputId, width = "100\%", height = "400px")

renderBpmnVisualization(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates an 'bpmnVisualization' HTML widget (or a promise of an 'bpmnVisualization' HTML widget).}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.}
}
\value{
An output or render function that enables the use of the 'bpmnVisualization' widget within Shiny applications.

An output or render function that enables the use of the 'bpmnVisualization' widget within Shiny applications.
}
\description{
Helpers to create output functions for using 'bpmnVisualization' HTML widgets within Shiny applications and interactive Rmd documents.

Helpers to create render functions for using 'bpmnVisualization' HTML widgets within Shiny applications and interactive Rmd documents.
}
