% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_fill}
\alias{create_fill}
\title{Create the fill style of an overlay}
\usage{
create_fill(color)
}
\arguments{
\item{color}{The color of the background of an overlay}
}
\value{
The fill style object of an overlay
}
\description{
When an overlay is added to an existing element in the diagram, it's possible to style how it is filled.

See \code{fill} argument in the \code{\link{create_style}} function.

Use this structure to create correct fill structure.
}
