% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_font}
\alias{create_font}
\title{Create the font style of an overlay}
\usage{
create_font(color = NULL, size = NULL)
}
\arguments{
\item{color}{The color of the font of the overlay}

\item{size}{The size of the font of the overlay}
}
\value{
The font style object of the overlay
}
\description{
When adding an overlay to an existing element in a diagram, it's possible to customize its font style.

Refer to the \code{font} parameter in the \code{\link{create_style}} function for more information.

Use this function to create the correct font structure for an overlay.
}
