% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmnVisualization.R
\name{create_overlay}
\alias{create_overlay}
\title{create_overlay}
\usage{
create_overlay(elementId, label)
}
\arguments{
\item{elementId}{The bpmn element id to which the overlay will be attached}

\item{label}{HTML element to use as an overlay to use as an overlay}
}
\value{
An overlay object
}
\description{
An overlay can be added to existing elements in the diagram.
See \code{overlays} argument in \code{\link{bpmnVisualization}} function.
Use this structure to create correct overlay structure.
}
