% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_overlay}
\alias{create_overlay}
\title{Create an overlay}
\usage{
create_overlay(elementId, label)
}
\arguments{
\item{elementId}{The bpmn element id to which the overlay will be attached}

\item{label}{HTML element to use as an overlay}
}
\value{
An overlay object
}
\description{
An overlay can be added to existing elements in the diagram.

See \code{overlays} argument in the \code{\link{display}} function.

Use this structure to create correct overlay structure.
}
