% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_stroke}
\alias{create_stroke}
\title{Create the stroke style of an overlay}
\usage{
create_stroke(color)
}
\arguments{
\item{color}{The color of the stroke of the overlay}
}
\value{
The stroke style object of the overlay
}
\description{
When adding an overlay to an existing element in a diagram, it's possible to customize its stroke. style.

Refer to the \code{stroke.} parameter in the \code{\link{create_style}} function for more information.

Use this function to create the correct stroke structure for an overlay.
}
