% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_style}
\alias{create_style}
\title{Create the style of an overlay}
\usage{
create_style(font = NULL, fill = NULL, stroke = NULL)
}
\arguments{
\item{font}{The font style of an overlay
Use \code{\link{create_font}} function to create the style of an overlay object.}

\item{fill}{The fill style of an overlay
Use \code{\link{create_fill}} function to create the style of an overlay object.}

\item{stroke}{The stroke style of an overlay
Use \code{\link{create_stroke}} function to create the style of an overlay object.}
}
\value{
The style object of an overlay
}
\description{
When an overlay is added to an existing element in the diagram, it's possible to style it.

See \code{style} argument in the \code{\link{create_overlay}} function.

Use this structure to create correct style structure.
}
