% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmnVisualization.R
\name{display}
\alias{display}
\title{display}
\usage{
display(
  bpmnXML,
  overlays = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{bpmnXML}{A file name or xml document or string in BPMN XML format}

\item{overlays}{An element or a list of elements to be added to the diagram's existing elements.
Use overlay function to create an overlay object with content and relative position.}

\item{width}{The width used to display the widget}

\item{height}{The height used to display the widget}

\item{elementId}{The id of the HTML element to enclose the widget}
}
\value{
A 'bpmnVisualization' htmlwidget that will intelligently print itself into HTML in a variety of contexts 
     including the R console, within R Markdown documents, and within Shiny output bindings.
}
\description{
Display BPMN diagram based on BPMN definition in XML format
}
\examples{
# Load the BPMN file
bpmn_file <- system.file("examples/Email_Voting.bpmn", package = "bpmnVisualization")

# Display the BPMN diagram
display(bpmn_file)

# Display the BPMN diagram with overlays
overlays <- list(create_overlay("bpmn_element_id_1", "42"), create_overlay("bpmn_element_id_2", "9"))
display(bpmn_file, overlays)

}
\seealso{
{[create_overlay()]}
}
