% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmnVisualization.R
\name{display}
\alias{display}
\title{display}
\usage{
display(
  bpmnXML,
  overlays = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{bpmnXML}{A file name or xml document or string in BPMN XML format}

\item{overlays}{A list of elements to be added to the diagram's existing elements.
Use overlay function to create an overlay object with content and relative position.}

\item{width}{The width used to display the widget}

\item{height}{The height used to display the widget}

\item{elementId}{The id of the HTML element to enclose the widget}
}
\description{
Display BPMN diagram based on BPMN definition in XML format
}
