% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\docType{data}
\name{overlay_shape_position}
\alias{overlay_shape_position}
\title{The overlay positions on \code{Shape}}
\format{
An object of class \code{character} of length 8.
}
\usage{
overlay_shape_position
}
\description{
To specify the position when creating an overlay object that will be attached to BPMN \code{Shape} elements in the diagram.
}
\details{
Use these constants as the \code{position} argument in the \code{\link{create_overlay}} function.

\strong{Positions}:
\itemize{
\item \code{top-left}
\item \code{top-right}
\item \code{top-center}
\item \code{bottom-left}
\item \code{bottom-right}
\item \code{bottom-center}
\item \code{middle-left}
\item \code{middle-right}
}
}
\examples{
# Create an overlay at the top-left corner of a shape
overlay <- create_overlay(elementId = 1, label = "My label", position = overlay_shape_position[1])

}
\seealso{
\code{\link{create_overlay}}
}
\keyword{datasets}
